Revert "Added plotting features, as well as additional file manipulation feat…"
This reverts commit 995df9b0c7718c7f41b6639c7fc81307502f2bae.
