+++ gausstrajlib.egg-info/SOURCES.txt
@@ -2,17 +2,9 @@ LICENSE.txt
 README.txt
 setup.py
 gausstrajlib/__init__.py
-gausstrajlib/conversions.py
-gausstrajlib/errors.py
-gausstrajlib/pdp.py
-gausstrajlib/reader copy.py
-gausstrajlib/reader.py
-gausstrajlib/test.py
-gausstrajlib/trajectory.py
 gausstrajlib/translate.py
-gausstrajlib/writer.py
 gausstrajlib.egg-info/PKG-INFO
 gausstrajlib.egg-info/SOURCES.txt
 gausstrajlib.egg-info/dependency_links.txt
 gausstrajlib.egg-info/top_level.txt
-scripts/gt-ucli
+scripts/gt-ucli.py
