--- scripts/comt.xyz
@@ -1,8 +1,0 @@
-6
--0.0477796
-C		 1.553074997450		 0.335873625694		-0.008821034249		-0.038072559809		-0.017296356404		-0.012154032602
-H		 1.397778874121		 1.441343756947		 0.024544313115		 0.004188741280		 0.005307600729		 0.004002570780
-H		 2.651071792388		 0.044250797742		 0.060900517739		 0.014456154966		-0.002003362472		 0.004669191349
-H		-1.132005309479		-0.108002086579		 0.025548567937		-0.013095085592		 0.006244642037		-0.000923992366
-O		 0.665869206718		-0.495906553543		-0.000208884866		 0.004032550848		 0.016867793663		 0.003626552367
-O		-2.014833564134		 0.157120394914		-0.000166829252		 0.028490198307		-0.009120317554		 0.000779710472
