--- scripts/sgdmldatasetnoecpt1.xyz
@@ -1,8 +1,0 @@
-6
-Energy=-29.803436978900468 Properties=species:S:1:pos:R:3:forces:R:3
-C	1.64274	0.35128	-0.00008	-69.06239	-63.66168	0.02053
-H	1.41575	1.48322	0.00097	-0.00657	-39.09711	-0.02612
-H	2.75336	0.03761	-0.00142	-39.72451	2.49552	0.03960
-O	0.70917	-0.51815	0.00022	69.55085	106.91567	-0.02597
-O	-2.12959	0.16929	-0.00016	15.62575	-6.62979	0.01053
-H	-1.18575	-0.16671	0.00051	23.61687	-0.02261	-0.01857
