--- scripts/sgdmldatasetnoecpt19996.xyz
@@ -1,8 +1,0 @@
-6
-Energy=-53.547328719443215 Properties=species:S:1:pos:R:3:forces:R:3
-C	-1.14937	-1.77579	0.10101	-14.66616	-0.39914	8.76739
-H	1.95874	2.34359	0.47472	9.23377	44.82868	-10.02942
-H	-0.05460	-1.80584	0.04458	-22.10509	2.07739	-11.24454
-O	-2.26085	-1.66294	-0.33771	36.79974	-1.81664	2.35185
-O	2.79665	2.83547	0.21144	-39.27748	-60.65450	19.13090
-H	3.27785	1.99734	0.28190	30.01522	15.96420	-8.97618
