--- scripts/sgdmldatasetnoecpt2meow.xyz
@@ -1,8 +1,0 @@
-6
-Energy=-29.803311477005614 Properties=species:S:1:pos:R:3:forces:R:3
-6	1.55217	0.33190	-0.00008	34.51843	31.84409	-0.00741
-1	1.33769	1.40145	0.00091	0.00535	19.53350	0.01197
-1	2.60155	0.03554	-0.00132	19.86824	-1.25055	-0.02050
-8	0.67006	-0.48958	0.00021	-34.76821	-53.45379	0.01175
-8	-2.01217	0.15996	-0.00015	-7.79065	3.30692	-0.00539
-1	-1.12035	-0.15752	0.00050	-11.83316	0.01983	0.00958
